+++ admin/partials/history.php
@@ -13,6 +13,13 @@ $history->prepare_items();
 						<form method="get" action="">
 							<input type="hidden" name="page" value="<?php echo esc_attr( $history->page ); ?>"/>
 							<input type="hidden" name="tab" value="<?php echo esc_attr( $history->tab ); ?>"/>
+							<p class="legend">
+								<?php echo esc_html__( 'Legend:', 'tmsm-aquatonic-course-booking' ); ?>
+								<span class="legend-upgrade"><?php echo esc_html__( 'Possible Allotment Upgrade', 'tmsm-aquatonic-course-booking' ); ?></span>
+								<span class="legend-cool"><?php echo esc_html__( 'Allotment Poorly Filled', 'tmsm-aquatonic-course-booking' ); ?></span>
+								<span class="legend-warning"><?php echo esc_html__( 'Allotment Medium Filled', 'tmsm-aquatonic-course-booking' ); ?></span>
+								<span class="legend-danger"><?php echo esc_html__( 'Allotment Very Filled', 'tmsm-aquatonic-course-booking' ); ?></span>
+							</p>
 							<p class="search-box">
 								<input type="search" placeholder="<?php echo esc_attr__( 'Course Date (MM/DD/YYYY)', 'tmsm-aquatonic-course-booking' ); ?>" name="search_datecourse" value="<?php
 								$search_datecourse = isset( $_REQUEST['search_datecourse'] ) ? esc_attr( wp_unslash( $_REQUEST['search_datecourse'] ) ) : '';
@@ -41,23 +48,30 @@ $history->prepare_items();
 								<?php
 								foreach($history->items as $history_item){
 									echo '<tr>';
-									;
 									foreach($history->get_columns() as $column_key => $column_name ){
 
 										$difference = $history_item[ 'courseallotment' ] - $history_item[ 'ongoingbookings' ];
 										$color = 'danger';
-										if($difference >= 7){
-											$color = 'warning';
+										$upgrade = 'upgradeno';
+
+										if ( $history_item['canstart'] > 10 ) {
+											$upgrade = 'upgradeyes';
 										}
-										if($difference >= 15){
-											$color = 'cool';
+										if ( $difference >= 7 ) {
+											$color   = 'warning';
+											$upgrade = 'upgradeno';
+										}
+										if ( $difference >= 15 ) {
+											$color   = 'cool';
+											$upgrade = 'upgradeno';
 										}
 
-										if( $column_key === 'ongoingbookings'){
-											echo '<td class="'.$column_key.' '.$column_key.'-'.$color.'">';
-										}
-										else{
-											echo '<td class="'.$column_key.'">';
+										if ( $column_key === 'ongoingbookings' ) {
+											echo '<td class="' . $column_key . ' ' . $column_key . '-' . $color . '">';
+										} elseif ( $column_key === 'canstart' ) {
+											echo '<td class="' . $column_key . ' ' . $column_key . '-' . $upgrade . '">';
+										} else {
+											echo '<td class="' . $column_key . '">';
 										}
 
 										echo $history->column_default($history_item, $column_key);