+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 * Update translations
 * Fix update contacts by ID
-* New feature: save dashboard past history
+* New feature: save dashboard past history, see logs in the "History" tab
 
 ### 1.9.0: November 16th, 2021
 * Google Pay Pass: don't display errors on frontend